# Extend bench bucket series with skip chunk test case