@yeya24 2020-10-28
Did a skip chunks benchmark again between the current master and https://github.com/thanos-io/thanos/commit/2a7590f69a5ace6014ddcc932788056935f201ba, that's the result. @bwplotka 

```
➜  thanos git:(test-ind) ✗ scripts/cmp.sh 1 0
Comparing '/home/yeya24/hub/gowork/thanos/scripts/results/BenchmarkBucketSeriesSkipChunks/1/2020-10-28-T11-49-55.bench.out' (c894acae22d34110993fc456a95038db88c6af1a BenchmarkBucketSeriesSkipChunks ./store 2h 2m /home/yeya24/hub/gowork/thanos/scripts/../pkg on core #3) with (new) '/home/yeya24/hub/gowork/thanos/scripts/results/BenchmarkBucketSeriesSkipChunks/0/2020-10-28-T11-16-20.bench.out' (c894acae22d34110993fc456a95038db88c6af1a BenchmarkBucketSeriesSkipChunks ./store 2h 2m /home/yeya24/hub/gowork/thanos/scripts/../pkg on core #3)
name                                                                  old time/op    new time/op    delta
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1of1000000              550ms ± 0%     423ms ± 0%  -23.05%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/10of1000000             553ms ± 0%     496ms ± 0%  -10.34%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1000000of1000000        4.20s ± 0%     3.39s ± 0%  -19.40%
BucketSeriesSkipChunks/100000SeriesWith100Samples/1of10000000           45.5ms ± 0%    38.6ms ± 0%  -15.16%
BucketSeriesSkipChunks/100000SeriesWith100Samples/100of10000000         44.1ms ± 0%    36.1ms ± 0%  -18.07%
BucketSeriesSkipChunks/100000SeriesWith100Samples/10000000of10000000     328ms ± 0%     276ms ± 0%  -15.92%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/1of10000000           5.14ms ± 0%    4.53ms ± 0%  -11.88%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/100of10000000         5.17ms ± 0%    4.13ms ± 0%  -20.11%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/10000000of10000000    86.2ms ± 0%    16.5ms ± 0%  -80.80%

name                                                                  old alloc/op   new alloc/op   delta
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1of1000000              125MB ± 0%      98MB ± 0%  -21.73%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/10of1000000             125MB ± 0%      98MB ± 0%  -21.69%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1000000of1000000       1.31GB ± 0%    1.00GB ± 0%  -23.69%
BucketSeriesSkipChunks/100000SeriesWith100Samples/1of10000000           11.2MB ± 0%     8.4MB ± 0%  -24.75%
BucketSeriesSkipChunks/100000SeriesWith100Samples/100of10000000         11.2MB ± 0%     8.4MB ± 0%  -24.41%
BucketSeriesSkipChunks/100000SeriesWith100Samples/10000000of10000000     126MB ± 0%      96MB ± 0%  -23.62%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/1of10000000           5.90MB ± 0%    4.34MB ± 0%  -26.44%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/100of10000000         5.90MB ± 0%    4.34MB ± 0%  -26.44%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/10000000of10000000    38.0MB ± 0%    17.3MB ± 0%  -54.37%

name                                                                  old allocs/op  new allocs/op  delta
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1of1000000              1.51M ± 0%     0.76M ± 0%  -49.68%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/10of1000000             1.51M ± 0%     0.76M ± 0%  -49.68%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1000000of1000000        12.1M ± 0%      8.0M ± 0%  -33.44%
BucketSeriesSkipChunks/100000SeriesWith100Samples/1of10000000             151k ± 0%       76k ± 0%  -49.66%
BucketSeriesSkipChunks/100000SeriesWith100Samples/100of10000000           151k ± 0%       76k ± 0%  -49.65%
BucketSeriesSkipChunks/100000SeriesWith100Samples/10000000of10000000     1.21M ± 0%     0.80M ± 0%  -33.44%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/1of10000000              231 ± 0%       195 ± 0%  -15.58%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/100of10000000            231 ± 0%       195 ± 0%  -15.58%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/10000000of10000000      170k ± 0%        1k ± 0%  -99.68%

```